feat: :new: encapsulation Map
封装Map文件
